+++ src/pages/post/index.js
@@ -1,4 +1,4 @@
-import PostLayout from "components/layouts/PostLayout";
+import PostLayout from "components/layouts/PostLayout/PostLayout";
 
 export default function PostPage() {
   return <PostLayout />;
